Dependency injection improvements for SesSnsEventMapper
