+++ src/ses-sns-event-mapper/interfaces/dynamic-object-mapper.ts
@@ -1,0 +1,7 @@
+import { MappingPair } from '@dynamic-mapper/mapper';
+import { MappingMembers } from '@dynamic-mapper/mapper/lib/interface';
+
+export interface IDynamicObjectMapperService<T, K> {
+  readonly mappingPair: MappingPair<T, K>;
+  createMap(source: T, config?: Partial<MappingMembers<T, K>>): K;
+}
