+++ src/ses-sns-event-mapper/interfaces/mapped-ses-sns-event.ts
