+++ src/ses-sns-event-mapper/interfaces/ses-sns-event-mapper.service.ts
@@ -1,0 +1,6 @@
+import { MappedSesSnsEvent } from './mapped-ses-sns-event';
+import { SesSnsEvent } from './ses-sns-event';
+
+export interface ISesSnsEventMapperService {
+  createMappedSeSSnsEvent(sesSnsEvent: SesSnsEvent): MappedSesSnsEvent;
+}
