+++ src/ses-sns-event-mapper/interfaces/ses-sns-event.ts
