--- src/ses-sns-event-mapper/repositories/ses-sns-event.repository.ts
@@ -1,6 +1,0 @@
-import { MappedSesSnsEvent } from '../interfaces/mapped-ses-sns-event.interface';
-import { SesSnsEvent } from '../interfaces/ses-sns-event.interface';
-
-export interface SesSnsEventRepository {
-  createMappedSeSSnsEvent(sesSnsEvent: SesSnsEvent): MappedSesSnsEvent;
-}
