+++ src/ses-sns-event-mapper/ses-sns-event-mapper.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Post, Body } from '@nestjs/common';
-import { SesSnsEventMapperService } from './ses-sns-event-mapper.service';
+import { SesSnsEventMapperService } from './providers/ses-sns-event-mapper.service';
 import { CreateSesSnsEventDto } from './dto/create-ses-sns-event.dto';
-import { MappedSesSnsEvent } from './interfaces/mapped-ses-sns-event.interface';
+import { MappedSesSnsEvent } from './interfaces/mapped-ses-sns-event';
 
 @Controller('ses-sns-event-mapper')
 export class SesSnsEventMapperController {
