+++ src/ses-sns-event-mapper/ses-sns-event-mapper.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
-import { SesSnsEventMapperService } from './ses-sns-event-mapper.service';
+import { SesSnsEventMapperService } from './providers/ses-sns-event-mapper.service';
 import { SesSnsEventMapperController } from './ses-sns-event-mapper.controller';
+import { DynamicObjectMapperService } from './providers/dynamic-object-mapper.service';
 
 @Module({
   controllers: [SesSnsEventMapperController],
-  providers: [SesSnsEventMapperService],
+  providers: [SesSnsEventMapperService, DynamicObjectMapperService],
 })
 export class SesSnsEventMapperModule {}
